feat(exam): set up to show alarm
set up UI and VM to show alarms on ExamScreen
split ExamScreen.kt
